refactor: split RunnerDaemon.swift (1,404 lines → 2 files), per #491
The WorkerDaemon actor's per-job processing pipeline (process →
executeTestSuites → interpretOutput → makeCollection) is the largest
contiguous block in the file.  Moved it to its own extension file so
the main file holds the lifecycle (run/workerLoop/sendHeartbeat) and
the subprocess/networking helpers.

Layout:
  RunnerDaemon.swift                — top types, WorkerCommand, secret
                                       resolution, actor decl + props +
                                       init, run + workerLoop, download,
                                       sendHeartbeat, connection-state
                                       tracking, runtime-helpers (893)
  RunnerDaemon+JobProcessing.swift  — process, executeTestSuites,
                                       interpretOutput, makeCollection (528)

Visibility changes (internal-only):
- Actor stored properties (poller, reporter, runner, …) dropped
  `private` so the extension can read them.
- `static let downloadSession` dropped `private`.
- Several file-private types/helpers used by the moved methods are now
  internal: JobStageTimings, RunnerJobStatus, ScriptResultJSON,
  TestStatus extension, mergeDirectoryContents,
  legacyPreferredStudentModuleFilename, shouldNormalizePythonSubmission,
  testSetupCacheKey, runMake, write*RuntimeHelper(s),
  writeStudentModuleHint, reportProcessingFailure, normalizedTestID,
  inferredCollectionStatus, recordConnectionLost/RestoredIfNeeded,
  process.

WorkerTests + ScriptInvocationTests pass.
