+++ Sources/Worker/RunnerDaemon.swift
@@ -8,14 +8,14 @@ import Foundation
 import FoundationNetworking  // URLSession, URLRequest on Linux
 #endif
 
-private enum RunnerJobStatus: String {
+enum RunnerJobStatus: String {
     case passed
     case failed
     case error
     case timeout
 }
 
-private struct JobStageTimings {
+struct JobStageTimings {
     private var values: [String: Int] = [:]
 
     mutating func measureSync<T>(_ stage: String, operation: () throws -> T) rethrows -> T {
@@ -248,26 +248,26 @@ func readWorkerSecretFromFile(path: String) -> String? {
 // MARK: - WorkerDaemon actor
 
 actor WorkerDaemon {
-    private static let downloadSession: URLSession = {
+    static let downloadSession: URLSession = {
         let cfg = URLSessionConfiguration.default
         cfg.timeoutIntervalForRequest = 5
         cfg.timeoutIntervalForResource = 15
         return URLSession(configuration: cfg)
     }()
 
-    private let poller: any JobPolling
-    private let reporter: any Reporting
-    private let runner: any ScriptRunner
-    private let apiBaseURL: URL
-    private let workerID: String
-    private let signer: WorkerRequestSigner
-    private let maxConcurrentJobs: Int
-    private let runnerProfile: RunnerCapabilityProfile?
-    private let downloadRetryPolicy: RunnerRetryPolicy
-    private let testSetupCache: TestSetupCache
-    private let config: RunnerDaemonConfig
-    private var serverConnectionLost = false
-    private var activeJobs = 0
+    let poller: any JobPolling
+    let reporter: any Reporting
+    let runner: any ScriptRunner
+    let apiBaseURL: URL
+    let workerID: String
+    let signer: WorkerRequestSigner
+    let maxConcurrentJobs: Int
+    let runnerProfile: RunnerCapabilityProfile?
+    let downloadRetryPolicy: RunnerRetryPolicy
+    let testSetupCache: TestSetupCache
+    let config: RunnerDaemonConfig
+    var serverConnectionLost = false
+    var activeJobs = 0
 
     init(
         poller: any JobPolling,
@@ -470,517 +470,6 @@ actor WorkerDaemon {
         }
     }
 
-    // MARK: - Job processing
-
-    private func process(_ job: Job) async throws {
-        activeJobs += 1
-        let jobStartedAt = Date()
-        defer { activeJobs = max(0, activeJobs - 1) }
-        var stageTimings = JobStageTimings()
-
-        writeStructuredRunnerLog(
-            event: "job_accepted",
-            fields: [
-                "runner_id": workerID,
-                "submission_id": job.submissionID,
-                "job_id": job.submissionID,
-                "test_setup_id": job.testSetupID,
-                "attempt_number": job.attemptNumber,
-                "runner_active_jobs": activeJobs,
-                "max_jobs": maxConcurrentJobs,
-            ])
-        try? await sendHeartbeat()
-
-        // Heartbeat loop scoped to the lifetime of this job. We use a manual
-        // Task + defer-cancel rather than a `withTaskGroup` because the
-        // remainder of `process()` is straight-line code with many local
-        // bindings; wrapping it in a group closure would balloon nesting
-        // without changing behaviour.
-        //
-        // Cancellation flow: when the outer worker loop is cancelled, the
-        // current `await` in `process()` throws, control runs to the defer,
-        // we cancel the heartbeat task explicitly, and `Task.sleep`'s
-        // `CancellationError` short-circuits the loop on its next wake.
-        let heartbeatTask = Task {
-            while !Task.isCancelled {
-                do {
-                    try await Task.sleep(for: .seconds(30))
-                } catch {
-                    break  // cancellation: skip the final heartbeat
-                }
-                try? await self.sendHeartbeat()
-            }
-        }
-        defer {
-            heartbeatTask.cancel()
-            // Fire-and-forget end-of-job heartbeat so the server's
-            // last-seen timestamp advances even if this job took >30s.
-            Task { try? await self.sendHeartbeat() }
-        }
-
-        let tempRoot = FileManager.default.temporaryDirectory
-        let freeDiskMBAtStart = freeSpaceMB(at: tempRoot)
-        if config.minFreeDiskMB > 0,
-            let freeMB = freeDiskMBAtStart,
-            freeMB < config.minFreeDiskMB
-        {
-            writeStructuredRunnerLog(
-                event: "insufficient_disk_space",
-                fields: [
-                    "runner_id": workerID,
-                    "submission_id": job.submissionID,
-                    "path": tempRoot.path,
-                    "free_mb": freeMB,
-                    "required_mb": config.minFreeDiskMB,
-                ])
-            throw WorkerDaemonError.insufficientDiskSpace(
-                path: tempRoot.path,
-                freeMB: freeMB,
-                requiredMB: config.minFreeDiskMB
-            )
-        }
-
-        let workDir =
-            tempRoot
-            .appendingPathComponent("chickadee_\(job.submissionID)_\(UUID().uuidString)", isDirectory: true)
-        let workDirSetupStartedAt = Date()
-        try FileManager.default.createDirectory(at: workDir, withIntermediateDirectories: true)
-        stageTimings.record("workdir_setup", milliseconds: Int(Date().timeIntervalSince(workDirSetupStartedAt) * 1000))
-
-        // Mutable so the body can populate them before building the
-        // diagnostics report (and so the defer can emit a structured event
-        // even on error paths where the body didn't reach the report).
-        // `freeDiskMBAtEnd` is sampled just before cleanup, so it represents
-        // the worst-case free-disk reading for this job.
-        var workdirPeakBytes: Int? = nil
-        var freeDiskMBAtEnd: Int? = nil
-        defer {
-            // If the happy path already measured these (right before the
-            // report), don't double-walk the directory.
-            if workdirPeakBytes == nil {
-                workdirPeakBytes = directorySizeBytes(at: workDir)
-            }
-            if freeDiskMBAtEnd == nil {
-                freeDiskMBAtEnd = freeSpaceMB(at: tempRoot)
-            }
-
-            let cleanupStartedAt = Date()
-            try? FileManager.default.removeItem(at: workDir)
-            stageTimings.record("cleanup", milliseconds: Int(Date().timeIntervalSince(cleanupStartedAt) * 1000))
-
-            let freeDiskMBPostCleanup = freeSpaceMB(at: tempRoot)
-            let totalWallClockMs = Int(Date().timeIntervalSince(jobStartedAt) * 1000)
-            var fields: [String: Any] = [
-                "runner_id": workerID,
-                "submission_id": job.submissionID,
-                "job_id": job.submissionID,
-                "total_wall_clock_ms": totalWallClockMs,
-            ]
-            for (key, value) in stageTimings.fields() {
-                fields[key] = value
-            }
-            writeStructuredRunnerLog(event: "job_stage_timings", fields: fields)
-
-            // Emit a dedicated disk-usage event so ops can answer "are we
-            // close to the floor?" without having to join across log events.
-            var diskFields: [String: Any] = [
-                "runner_id": workerID,
-                "submission_id": job.submissionID,
-                "job_id": job.submissionID,
-                "path": tempRoot.path,
-                "min_free_disk_mb": config.minFreeDiskMB,
-            ]
-            if let v = freeDiskMBAtStart { diskFields["free_disk_mb_at_start"] = v }
-            if let v = freeDiskMBAtEnd { diskFields["free_disk_mb_at_end"] = v }
-            if let v = freeDiskMBPostCleanup { diskFields["free_disk_mb_post_cleanup"] = v }
-            if let v = workdirPeakBytes { diskFields["workdir_peak_bytes"] = v }
-            writeStructuredRunnerLog(event: "job_disk_usage", fields: diskFields)
-        }
-
-        // Download submission and acquire the prepared test setup concurrently.
-        // The test setup is served from the LRU cache: on a hit the cached
-        // directory is copied into a fresh scratch location; on a miss it is
-        // downloaded, unzipped, committed to cache, then copied.
-        let submissionZip = workDir.appendingPathComponent("submission.zip")
-        let submissionDir = workDir.appendingPathComponent("submission", isDirectory: true)
-        try stageTimings.measureSync("submission_dir_setup") {
-            try FileManager.default.createDirectory(at: submissionDir, withIntermediateDirectories: true)
-        }
-
-        let submissionDownloadStartedAt = Date()
-        async let submissionDownload: Void = download(url: job.submissionURL, to: submissionZip)
-
-        let testSetupAcquireStartedAt = Date()
-        let testSetupCacheKey = testSetupCacheKey(for: job)
-        let testSetupDir = try await testSetupCache.acquire(testSetupID: testSetupCacheKey) {
-            let stagingZip = workDir.appendingPathComponent("testsetup.zip")
-            let stagingDir = workDir.appendingPathComponent("testsetup_staging", isDirectory: true)
-            try FileManager.default.createDirectory(at: stagingDir, withIntermediateDirectories: true)
-            try await self.download(url: job.testSetupURL, to: stagingZip)
-            try self.unzip(stagingZip, to: stagingDir)
-            return stagingDir
-        }
-        stageTimings.record(
-            "test_setup_acquire", milliseconds: Int(Date().timeIntervalSince(testSetupAcquireStartedAt) * 1000))
-        defer { try? FileManager.default.removeItem(at: testSetupDir) }
-
-        try await submissionDownload
-        stageTimings.record(
-            "submission_download", milliseconds: Int(Date().timeIntervalSince(submissionDownloadStartedAt) * 1000))
-
-        let manifest = job.manifest
-
-        // Stage the submission independently from the grading workspace so the
-        // worker can normalize it without mutating the raw artifact.
-        try stageTimings.measureSync("submission_unpack") {
-            if let filename = job.submissionFilename {
-                let dest = stagedSubmissionDestination(
-                    submissionDirectory: submissionDir,
-                    submittedFilename: filename
-                )
-                try FileManager.default.createDirectory(
-                    at: dest.deletingLastPathComponent(),
-                    withIntermediateDirectories: true
-                )
-                try? FileManager.default.removeItem(at: dest)
-                try FileManager.default.copyItem(at: submissionZip, to: dest)
-            } else {
-                try unzip(submissionZip, to: submissionDir)
-            }
-        }
-
-        // Remove the starter notebook template from the test directory so
-        // grading scripts that scan for *.ipynb don't see both the template
-        // and the student/canonical submission.  Older manifests lack
-        // starterNotebook — fall back to "assignment.ipynb" since that is
-        // the conventional name used by every existing assignment.
-        try stageTimings.measureSync("starter_cleanup") {
-            let starterName = manifest.starterNotebook ?? "assignment.ipynb"
-            do {
-                let starterPath = testSetupDir.appendingPathComponent(starterName)
-                if FileManager.default.fileExists(atPath: starterPath.path),
-                    job.submissionFilename != starterName
-                {
-                    try FileManager.default.removeItem(at: starterPath)
-                }
-            }
-        }
-
-        let normalizationWarnings: [String]
-        let preferredStudentModule: String?
-        (normalizationWarnings, preferredStudentModule) = try stageTimings.measureSync("submission_prepare") {
-            if shouldNormalizePythonSubmission(
-                manifest: manifest,
-                submissionFilename: job.submissionFilename,
-                submissionDirectory: submissionDir
-            ) {
-                let normalizer = SubmissionNormalizer()
-                let normalization = try normalizer.normalizePythonSubmission(
-                    manifest: manifest,
-                    submissionDirectory: submissionDir,
-                    workspaceDirectory: testSetupDir,
-                    submissionFilename: job.submissionFilename
-                )
-                return (normalization.warnings, normalization.preferredStudentModule)
-            } else {
-                try mergeDirectoryContents(from: submissionDir, into: testSetupDir)
-                try extractNotebooksToCode(in: testSetupDir)
-                return ([], legacyPreferredStudentModuleFilename(submissionFilename: job.submissionFilename))
-            }
-        }
-
-        // Optional make step.
-        try stageTimings.measureSync("make_step") {
-            if let makefile = manifest.makefile {
-                try runMake(in: testSetupDir, target: makefile.target)
-            }
-        }
-
-        // Install shared Python test runtime helpers for every run.
-        try stageTimings.measureSync("runtime_helper_setup") {
-            try writePythonRuntimeHelpers(in: testSetupDir)
-            try writeStudentModuleHint(in: testSetupDir, preferredFilename: preferredStudentModule)
-            try writeRRuntimeHelper(in: testSetupDir)
-        }
-
-        let testExecutionStartedAt = Date()
-        let outcomes = await executeTestSuites(
-            manifest: manifest,
-            testSetupDir: testSetupDir,
-            job: job
-        )
-        stageTimings.record(
-            "test_execution", milliseconds: Int(Date().timeIntervalSince(testExecutionStartedAt) * 1000))
-
-        // Sample disk usage at end-of-execution, before the report is sent,
-        // so the persisted diagnostics reflect this job's actual footprint.
-        // The defer will re-use these readings instead of walking again.
-        workdirPeakBytes = directorySizeBytes(at: workDir)
-        freeDiskMBAtEnd = freeSpaceMB(at: tempRoot)
-
-        let collection = makeCollection(
-            outcomes: outcomes,
-            warnings: normalizationWarnings,
-            job: job,
-            startedAt: jobStartedAt
-        )
-        let diagnostics = WorkerExecutionDiagnostics(
-            runnerID: workerID,
-            startedAt: jobStartedAt,
-            finishedAt: collection.timestamp,
-            finalStatus: inferredCollectionStatus(collection).rawValue,
-            timedOut: collection.timeoutCount > 0,
-            exitCode: nil,
-            terminationReason: nil,
-            peakRSSBytes: nil,
-            wallClockMs: collection.executionTimeMs,
-            childProcessCount: nil,
-            stdoutBytes: nil,
-            stderrBytes: nil,
-            stageTimings: stageTimings.asWorkerExecutionStageTimings(),
-            freeDiskMBAtStart: freeDiskMBAtStart,
-            freeDiskMBAtEnd: freeDiskMBAtEnd,
-            workdirPeakBytes: workdirPeakBytes
-        )
-        do {
-            let resultReportStartedAt = Date()
-            try await reporter.report(WorkerExecutionReport(collection: collection, diagnostics: diagnostics))
-            stageTimings.record(
-                "result_report", milliseconds: Int(Date().timeIntervalSince(resultReportStartedAt) * 1000))
-            writeStructuredRunnerLog(
-                event: "result_submission_succeeded",
-                fields: [
-                    "runner_id": workerID,
-                    "submission_id": job.submissionID,
-                    "status": inferredCollectionStatus(collection).rawValue,
-                ])
-        } catch {
-            writeStructuredRunnerLog(
-                event: "result_submission_failed",
-                fields: [
-                    "runner_id": workerID,
-                    "submission_id": job.submissionID,
-                    "error_type": String(describing: type(of: error)),
-                    "error_message_summary": String(describing: error),
-                ])
-            throw error
-        }
-    }
-
-    // MARK: - Test execution
-
-    /// Walks `manifest.testSuites` in order, honouring the `dependsOn`
-    /// pass-gate: a test whose prerequisite hasn't passed is auto-failed
-    /// with a `Skipped:` short result instead of executed. Missing script
-    /// files are logged and skipped entirely (no outcome emitted, matching
-    /// the pre-extraction behaviour).
-    private func executeTestSuites(
-        manifest: TestProperties,
-        testSetupDir: URL,
-        job: Job
-    ) async -> [TestOutcome] {
-        var outcomes: [TestOutcome] = []
-        var passedScripts: Set<String> = []
-
-        for entry in manifest.testSuites {
-            if let blockedBy = entry.dependsOn.first(where: { !passedScripts.contains($0) }),
-                !entry.dependsOn.isEmpty
-            {
-                let baseName = (entry.script as NSString).deletingPathExtension
-                let displayName = entry.name.flatMap { $0.trimmingCharacters(in: .whitespaces).isEmpty ? nil : $0 }
-                outcomes.append(
-                    TestOutcome(
-                        testName: displayName ?? (baseName.isEmpty ? entry.script : baseName),
-                        testClass: nil,
-                        tier: entry.tier,
-                        status: .fail,
-                        shortResult: "Skipped: prerequisite '\(blockedBy)' did not pass",
-                        longResult: nil,
-                        executionTimeMs: 0,
-                        memoryUsageBytes: nil,
-                        attemptNumber: job.attemptNumber,
-                        isFirstPassSuccess: false
-                    ))
-                continue
-            }
-
-            let scriptURL = testSetupDir.appendingPathComponent(entry.script)
-            guard FileManager.default.fileExists(atPath: scriptURL.path) else {
-                writeStructuredRunnerLog(
-                    event: "local_execution_error",
-                    fields: [
-                        "runner_id": workerID,
-                        "submission_id": job.submissionID,
-                        "test_id": entry.script,
-                        "error_type": "missing_script",
-                        "error_message_summary": entry.script,
-                    ])
-                continue
-            }
-
-            writeStructuredRunnerLog(
-                event: "test_execution_start",
-                fields: [
-                    "runner_id": workerID,
-                    "submission_id": job.submissionID,
-                    "test_id": entry.script,
-                ])
-
-            // Phase 1 of issue #461 — surface the per-(student, assignment)
-            // seed to the grading subprocess. Nil seed means non-personalized
-            // job; leaving the env var unset preserves legacy behaviour.
-            var scriptEnv: [String: String] = [:]
-            if let seed = job.assignmentSeed, !seed.isEmpty {
-                scriptEnv["CHICKADEE_ASSIGNMENT_SEED"] = seed
-            }
-            let output = await runner.run(
-                script: scriptURL,
-                workDir: testSetupDir,
-                timeLimitSeconds: manifest.timeLimitSeconds,
-                env: scriptEnv
-            )
-
-            let isFirstAttempt = job.attemptNumber == 1
-            let outcome = interpretOutput(
-                output, entry: entry, attemptNumber: job.attemptNumber, isFirstAttempt: isFirstAttempt)
-            outcomes.append(outcome)
-            writeStructuredRunnerLog(
-                event: output.timedOut ? "timeout" : "test_execution_end",
-                fields: [
-                    "runner_id": workerID,
-                    "submission_id": job.submissionID,
-                    "test_id": normalizedTestID(for: outcome),
-                    "status": outcome.status.rawValue,
-                    "execution_ms": outcome.executionTimeMs,
-                ])
-            if outcome.status == .pass {
-                passedScripts.insert(entry.script)
-            }
-        }
-
-        return outcomes
-    }
-
-    // MARK: - Script output interpretation
-
-    private func interpretOutput(
-        _ output: ScriptOutput,
-        entry: TestSuiteEntry,
-        attemptNumber: Int,
-        isFirstAttempt: Bool
-    ) -> TestOutcome {
-        let status: TestStatus
-        if output.timedOut {
-            status = .timeout
-        } else {
-            switch output.exitCode {
-            case 0: status = .pass
-            case 1: status = .fail
-            case 3: status = .fail  // chickadee.py (Marmoset) uses exit 3 for "failed"
-            default: status = .error
-            }
-        }
-
-        // Parse the last non-empty stdout line as optional JSON for score/shortResult.
-        let lastLine = output.stdout
-            .components(separatedBy: "\n")
-            .map { $0.trimmingCharacters(in: .whitespaces) }
-            .last(where: { !$0.isEmpty })
-
-        var shortResult: String
-
-        if let line = lastLine,
-            let data = line.data(using: .utf8),
-            let json = try? JSONDecoder().decode(ScriptResultJSON.self, from: data)
-        {
-            shortResult = json.shortResult ?? status.defaultShortResult
-            // json.score reserved for Phase 5 gamification
-        } else if let line = lastLine {
-            shortResult = line
-        } else {
-            shortResult = status.defaultShortResult
-        }
-
-        let stderrText = output.stderr.trimmingCharacters(in: .whitespacesAndNewlines)
-        // Strip the JSON footer line from stdout before displaying to students.
-        // The footer is the last non-empty line; if it parsed as JSON above we
-        // remove it so only human-readable output appears in longResult.
-        let strippedStdout: String = {
-            guard let line = lastLine,
-                let data = line.data(using: .utf8),
-                (try? JSONDecoder().decode(ScriptResultJSON.self, from: data)) != nil
-            else { return output.stdout }
-            var lines = output.stdout.components(separatedBy: "\n")
-            if let lastIdx = lines.indices.last(where: { !lines[$0].trimmingCharacters(in: .whitespaces).isEmpty }) {
-                lines.remove(at: lastIdx)
-            }
-            return lines.joined(separator: "\n")
-        }()
-        let stdoutText = strippedStdout.trimmingCharacters(in: .whitespacesAndNewlines)
-        let longResult: String? = {
-            var sections: [String] = []
-            if !stdoutText.isEmpty { sections.append("stdout:\n\(stdoutText)") }
-            if !stderrText.isEmpty { sections.append("stderr:\n\(stderrText)") }
-            return sections.isEmpty ? nil : sections.joined(separator: "\n\n")
-        }()
-        let baseName = (entry.script as NSString).deletingPathExtension
-        let displayName = entry.name.flatMap { $0.trimmingCharacters(in: .whitespaces).isEmpty ? nil : $0 }
-
-        return TestOutcome(
-            testName: displayName ?? (baseName.isEmpty ? entry.script : baseName),
-            testClass: nil,
-            tier: entry.tier,
-            status: status,
-            shortResult: shortResult,
-            longResult: longResult,
-            points: entry.points,
-            executionTimeMs: output.executionTimeMs,
-            memoryUsageBytes: nil,
-            attemptNumber: attemptNumber,
-            isFirstPassSuccess: isFirstAttempt && status == .pass
-        )
-    }
-
-    // MARK: - Collection assembly
-
-    private func makeCollection(
-        outcomes: [TestOutcome],
-        warnings: [String],
-        job: Job,
-        startedAt: Date
-    ) -> TestOutcomeCollection {
-        let passCount = outcomes.filter { $0.status == .pass }.count
-        let failCount = outcomes.filter { $0.status == .fail }.count
-        let errorCount = outcomes.filter { $0.status == .error }.count
-        let timeoutCount = outcomes.filter { $0.status == .timeout }.count
-        let totalMs = outcomes.reduce(0) { $0 + $1.executionTimeMs }
-        let totalPoints = outcomes.reduce(0) { $0 + $1.points }
-        let earnedPoints = outcomes.filter { $0.status == .pass }.reduce(0) { $0 + $1.points }
-
-        let buildStatus: BuildStatus = outcomes.isEmpty ? .failed : .passed
-
-        return TestOutcomeCollection(
-            submissionID: job.submissionID,
-            testSetupID: job.testSetupID,
-            attemptNumber: job.attemptNumber,
-            buildStatus: buildStatus,
-            compilerOutput: nil,
-            outcomes: outcomes,
-            totalTests: outcomes.count,
-            passCount: passCount,
-            failCount: failCount,
-            errorCount: errorCount,
-            timeoutCount: timeoutCount,
-            executionTimeMs: totalMs,
-            totalPoints: totalPoints,
-            earnedPoints: earnedPoints,
-            warnings: warnings,
-            jobStartedAt: startedAt,
-            runnerVersion: ChickadeeVersion.current,
-            timestamp: Date()
-        )
-    }
-
     // MARK: - Subprocess helpers
 
     func download(url: URL, to destination: URL) async throws {
@@ -1056,7 +545,7 @@ actor WorkerDaemon {
         }
     }
 
-    private func runMake(in directory: URL, target: String?) throws {
+    func runMake(in directory: URL, target: String?) throws {
         let proc = Process()
         proc.executableURL = URL(fileURLWithPath: "/usr/bin/make")
         proc.arguments = target.map { [$0] } ?? []
@@ -1070,7 +559,7 @@ actor WorkerDaemon {
 
     // extractNotebooksToCode is a module-level function (NotebookExtractor.swift).
 
-    private func reportProcessingFailure(job: Job, error: Error) async throws {
+    func reportProcessingFailure(job: Job, error: Error) async throws {
         let message = String(describing: error)
         let collection = TestOutcomeCollection(
             submissionID: job.submissionID,
@@ -1115,7 +604,7 @@ actor WorkerDaemon {
         }
     }
 
-    private func recordConnectionLostIfNeeded(
+    func recordConnectionLostIfNeeded(
         stage: RunnerRetryStage,
         message: String,
         retryInSeconds: Int?
@@ -1144,7 +633,7 @@ actor WorkerDaemon {
         }
     }
 
-    private func recordConnectionRestoredIfNeeded(stage: RunnerRetryStage) {
+    func recordConnectionRestoredIfNeeded(stage: RunnerRetryStage) {
         guard serverConnectionLost else { return }
         serverConnectionLost = false
         writeStructuredRunnerLog(
@@ -1156,24 +645,24 @@ actor WorkerDaemon {
             ])
     }
 
-    private func normalizedTestID(for outcome: TestOutcome) -> String {
+    func normalizedTestID(for outcome: TestOutcome) -> String {
         let classPart = outcome.testClass?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
         return classPart.isEmpty ? outcome.testName : "\(classPart).\(outcome.testName)"
     }
 
-    private func inferredCollectionStatus(_ collection: TestOutcomeCollection) -> RunnerJobStatus {
+    func inferredCollectionStatus(_ collection: TestOutcomeCollection) -> RunnerJobStatus {
         if collection.timeoutCount > 0 { return .timeout }
         if collection.errorCount > 0 { return .error }
         if collection.buildStatus == .failed || collection.failCount > 0 { return .failed }
         return .passed
     }
 
-    private func writeRRuntimeHelper(in directory: URL) throws {
+    func writeRRuntimeHelper(in directory: URL) throws {
         let rRuntimeURL = directory.appendingPathComponent("test_runtime.R")
         try testRuntimeR.write(to: rRuntimeURL, atomically: true, encoding: .utf8)
     }
 
-    private func writePythonRuntimeHelpers(in directory: URL) throws {
+    func writePythonRuntimeHelpers(in directory: URL) throws {
         let runtimeURL = directory.appendingPathComponent("test_runtime.py")
         try testRuntimePy.write(to: runtimeURL, atomically: true, encoding: .utf8)
 
@@ -1183,7 +672,7 @@ actor WorkerDaemon {
         try sitecustomizePy.write(to: sitecustomizeURL, atomically: true, encoding: .utf8)
     }
 
-    private func writeStudentModuleHint(in directory: URL, preferredFilename: String?) throws {
+    func writeStudentModuleHint(in directory: URL, preferredFilename: String?) throws {
         let hintURL = directory.appendingPathComponent(".chickadee_student_module")
         try? FileManager.default.removeItem(at: hintURL)
 
@@ -1237,7 +726,7 @@ func directorySizeBytes(at directory: URL) -> Int? {
     return total
 }
 
-private func mergeDirectoryContents(from sourceDirectory: URL, into destinationDirectory: URL) throws {
+func mergeDirectoryContents(from sourceDirectory: URL, into destinationDirectory: URL) throws {
     // Resolve symlinks once on the source root so that the prefix comparison
     // below works even when callers pass paths through `/var` vs `/private/var`
     // (macOS) or otherwise-aliased mounts.
@@ -1282,7 +771,7 @@ private func mergeDirectoryContents(from sourceDirectory: URL, into destinationD
     }
 }
 
-private func legacyPreferredStudentModuleFilename(submissionFilename: String?) -> String? {
+func legacyPreferredStudentModuleFilename(submissionFilename: String?) -> String? {
     guard let submissionFilename, !submissionFilename.isEmpty else { return nil }
     let submittedName = URL(fileURLWithPath: submissionFilename).lastPathComponent
     guard !submittedName.isEmpty else { return nil }
@@ -1306,7 +795,7 @@ func stagedSubmissionDestination(
     return submissionDirectory.appendingPathComponent(safeName)
 }
 
-private func shouldNormalizePythonSubmission(
+func shouldNormalizePythonSubmission(
     manifest: TestProperties,
     submissionFilename: String?,
     submissionDirectory: URL
@@ -1347,7 +836,7 @@ private func shouldNormalizePythonSubmission(
     return false
 }
 
-private func testSetupCacheKey(for job: Job) -> String {
+func testSetupCacheKey(for job: Job) -> String {
     let manifestBytes = (try? ManifestCodec.encoder.encode(job.manifest)) ?? Data()
     var material = Data()
     material.append(Data(job.testSetupID.utf8))
@@ -1362,14 +851,14 @@ private func testSetupCacheKey(for job: Job) -> String {
 // MARK: - Script result JSON (optional last-line protocol)
 
 /// Scripts may optionally write this as their last stdout line to report a score.
-private struct ScriptResultJSON: Decodable {
+struct ScriptResultJSON: Decodable {
     let score: Double?
     let shortResult: String?
 }
 
 // MARK: - Helpers
 
-private extension TestStatus {
+extension TestStatus {
     var defaultShortResult: String {
         switch self {
         case .pass: return "passed"
